Fix: `/manifest.json` returning 404
After creating react app with `bun create react` and starting dev server, server is responding with a status of 404 because it was reading /manifest.json as `application/octet-stream` instead of `application/json`. In my WSL (Ubuntu 20.04) server responds with this message after running dev server - `404 GET /manifest.json as application/octet-stream`. So, I assume this is because these 'if' statements are not placed correctly, and should be placed inside `if (strings.eqlComptimeIgnoreLen(category_, "application")) {` condition.